+++ releasenotes.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Version 0.26.0
+### New Features
+- Date.Forward(int days = 365)
+
 ## Version 0.25.0
 ### New Features
 - Date.Between(DateTime minDate, DateTime maxDate)